requirements.txt
streamlit
python-dotenv
crewai
pip install crewai crewai_tools
pandas
plotly
matplotlib
seaborn
fpdf
langchain_groq
faiss-cpu